Fixed some issues in the model. - Removed overlapping/duplicate pieces - Moved things to be closer to one another and more symetrical - Changed slicer settings to only have 3 layers for the wall, which I think will help for the issue where there are random gaps. - Also removed the beveled corners in the base of the original model
Lastly, I didn't say this last time but one of the issues I had with the printer seemed it was too hot and the fan wasn't running as much? I am not entirely sure. But after setting the temp to 220 and setting the fan to run at all times things seem to work better.